+++ R/find_algorithm.R
@@ -95,6 +95,12 @@ find_algorithm.BBreg <- function(x, ...) {
 
 
 #' @export
+find_algorithm.glimML <- function(x, ...) {
+  list("algorithm" = "ML")
+}
+
+
+#' @export
 find_algorithm.BBmm <- function(x, ...) {
   method <- parse(text = deparse(x$call, width.cutoff = 500))[[1]]$method
   if (is.null(method)) method <- "BB-NR"