+++ R/find_parameters.R
@@ -179,6 +179,21 @@ find_parameters.BBmm <- function(x, flatten = FALSE, ...) {
 
 
 #' @export
+find_parameters.glimML <- function(x, flatten = FALSE, ...) {
+  l <- compact_list(list(
+    conditional = names(x@fixed.param),
+    random = names(x@random.param)
+  ))
+
+  if (flatten) {
+    unique(unlist(l))
+  } else {
+    l
+  }
+}
+
+
+#' @export
 find_parameters.lrm <- function(x, flatten = FALSE, ...) {
   l <- list(conditional = names(stats::coef(x)))
   if (flatten) {